Execution with node {file name} which would seed the db rather than having an endpoint exposed on the server that triggers these methods
