+++ packages/EntityBackend/server.js
@@ -1,12 +1,14 @@
 // server.js
-const config = require('./config');
-const entityRouter = require('./routes/entityRouter');
+const config = require("./config");
+const entityRouter = require("./routes/entityRouter");
 const express = require("express");
-const cors = require('cors')
+const cors = require("cors");
 const server = express();
+const GenerateData = require("./Helper/GenerateData");
 
 const body_parser = require("body-parser");
 
+GenerateData();
 // parse JSON (application/json content-type)
 server.use(body_parser.json());
 
@@ -16,5 +18,5 @@ server.use(cors());
 server.use("/", entityRouter);
 
 server.listen(port, () => {
-    console.log(`Server listening at ${port}`);
+  console.log(`Server listening at ${port}`);
 });
